refactored primites module
