+++ primitives/authenticated_encryption.py
@@ -1,5 +1,6 @@
 import os
 from cryptography.hazmat.primitives.ciphers.aead import AESGCM
+from utils import  bytes_to_hex_string, hex_string_to_bytes, to_bytes_like
 
 def generate_key(self, key_length):
     '''
@@ -16,7 +17,7 @@ def generate_key(self, key_length):
 
     key = AESGCM.generate_key(bit_length=key_length)
 
-    return _bytes_to_hex_string(key)
+    return bytes_to_hex_string(key)
 
 def AESGCM_encrypt(self, key, secret, unencrypted_data = None):
     '''
@@ -33,15 +34,15 @@ def AESGCM_encrypt(self, key, secret, unencrypted_data = None):
     if len(key) != 64:
         raise ValueError("The key must be 256 bit length")
     
-    bkey = _hex_string_to_bytes(key)
-    bsecret = _to_bytes_like(secret)
-    bud = _to_bytes_like(unencrypted_data) if unencrypted_data else None
+    bkey = hex_string_to_bytes(key)
+    bsecret = to_bytes_like(secret)
+    bud = to_bytes_like(unencrypted_data) if unencrypted_data else None
     
     aesgcm = AESGCM(bkey)
     nonce = os.urandom(12)
     bencrypted = aesgcm.encrypt(nonce, bsecret, bud)
 
-    return _bytes_to_hex_string(bencrypted)
+    return bytes_to_hex_string(bencrypted)
 
 def AESGCM_decrypt(self, key):
     '''
@@ -54,49 +55,3 @@ def AESGCM_decrypt(self, key):
         string: the hash as hexadecimal string  
     '''
     return "to do"
-
-
-def _bytes_to_hex_string(byte_string):
-    '''
-        This function converts a bytes data to hexadecimal string
-
-        Parameters:
-        byte_string (byte): The data to convert
-
-        Returns:
-        string: the data converted to hexadecimal string  
-    '''
-    hex_string = byte_string.hex()
-    return hex_string
-
-def _hex_string_to_bytes(hex_string):
-    '''
-        This function converts a hexadecimal string data to bytes
-
-        Parameters:
-        hex_string (string): The data to convert
-
-        Returns:
-        bytes: the data converted to bytes  
-    '''
-    byte_key = bytes.fromhex(hex_string)
-    return byte_key
-
-def _to_bytes_like(data):
-    '''
-        This function converts a string, a list or a tuple in bytes-like data
-
-        Parameters:
-        data (string|list|tuple): The data to convert
-
-        Returns:
-        bytes: the data converted to bytes  
-    '''
-    if isinstance(data, str):
-        byte_data = data.encode()
-    elif isinstance(data, (list, tuple)):
-        byte_data = bytes(data)
-    else:
-        raise TypeError("The data type is not supported. Supported types are: string, list and tuple")
-    
-    return byte_data
