+++ primitives/test_authenticated_encryption.py
@@ -1,5 +1,5 @@
 import pytest
-from authenticated_encryption import generate_key, _bytes_to_hex_string, AESGCM_encrypt
+from authenticated_encryption import generate_key, AESGCM_encrypt
 
 def test_generate_key_valid_lengths():
     for key_length in [128, 192, 256]:
@@ -9,12 +9,6 @@ def test_generate_key_valid_lengths():
 def test_generate_key_invalid_length():
     with pytest.raises(ValueError):
         generate_key("", 100)
-
-def test_bytes_to_hex_string():
-    byte_key = b'\x00\x01\x02\x03\x04\x05'
-    expected_hex_string = '000102030405'
-    hex_string = _bytes_to_hex_string(byte_key)
-    assert hex_string == expected_hex_string
 
 def test_AESGCM_encrypt_key_length():
     key = '000102030405060708090a0b0c0d0e0f'
